requirements.txt
fastapi
uvicorn
python-dotenv
pillow
pyzbar
requests
python-multipart
python-whois==0.8.0
dnspython==2.4.2
